refactor: extract TempGitRepo test helper into shared test_utils module
Move the TempGitRepo helper from timeline.rs tests into a dedicated
src/test_utils.rs module so it can be reused by future test modules
without duplication.
